Wait for cluster ControlPlaneReady instead of Ready in bootstrap.sh
It may take a while for cluster to be ready in large deployments, which may
cause a timeout waiting for that condition.

In that case pivot may happen in the meantime, resulting in
management-cluster-kubeconfig secret transfert move to management-cluster

Waiting for ControlPlaneReady condition allows us to retrieve and use
management-cluster-kubeconfig as soon as possible.