telco-cloud-init chart: gotpl all chart values, preserve type and enable nested templating
Chart values templating has been introduced recently in telco-cloud-init chart,
it provides a very convenient way to reuse variable and introduce logic to
decide of their values.

Due to limitations of the 'tpl' function of the templating language, current
capabilities were constrained to produce string output. We introduce a
"preserve-type" filter that enables to preserve the type of templates that
produces values that are not strings (like interger, list, dicts)

We also apply templating on whole chart value, this way we won't have to enble
interpretation on various fields of the helm chart, which will help keeping
charts more readable, and will save us from having to document which fields
are interpreted

One (nice) side effect of this change, is that it eases the interpretation of
nested templates, that can now be supported (with a fixed/limited amount
of indirection)