Support existing management cluster
For a commercial use case where a production management cluster is
already deployed, there's no need to perform bootstrapping and pivot.
In this case, we only need to run `apply.sh` with a customized
`values.yaml` to ensure the required Sylva units are working in
harmony with the existing cluster.

To make this work, we need to remove the following constraints:

1. `management-cluster-kubeconfig` should not be mandatory as `sylvactl`
   is capable of working with the default kubeconfig (in
   ~/.kube/config).
2. `cluster` unit should not be mandatory. This essentially renders
   the `enabled` flag in the `cluster` unit ineffective. We should
   let the unit's enabled flag to determine whether the given
   unit should be enabled, for consistency.