transform 'workload-cluster-sriov*' units into new 'sriov*' units
The goal of this commit series is to ultimately have "sriov-crd",
"sriov" and "sriov-resources" be units that can be deployed on any cluster deployed via
sylva-units, either the mgmt cluster or a workload cluster.

This commit temporarily makes these units only deployable in mgmt cluster.

A commit later in this commit series is enabling these units in
workload-cluster.values.yaml to make them usable in any cluster, mgmt
cluster and workload cluster.